Show seed only on successful authentication
